+++ app/build.gradle.kts
@@ -2,6 +2,8 @@ plugins {
     alias(libs.plugins.android.application)
     alias(libs.plugins.kotlin.android)
     alias(libs.plugins.kotlin.compose)
+    /*alias(libs.plugins.hilt)
+    alias(libs.plugins.kotlin.ksp)*/
 }
 
 android {
@@ -53,6 +55,9 @@ dependencies {
 
     implementation(libs.coil)
     implementation(libs.retrofit)
+
+    /*implementation(libs.hilt.android)
+    ksp(libs.hilt.compiler)*/
