+++ app/src/main/java/com/artem/animationjikan/AnimationJikanApp.kt
@@ -1,0 +1,4 @@
+package com.artem.animationjikan
+
+class AnimationJikanApp {
+}
