+++ app/src/main/java/com/artem/animationjikan/ui/MainActivity.kt
@@ -1,0 +1,26 @@
+package com.artem.animationjikan.ui
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.compose.runtime.Composable
+import com.artem.animationjikan.ui.theme.AnimationJikanTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            MyApp()
+        }
+    }
+}
+
+
+@Composable
+private fun MyApp() {
+    AnimationJikanTheme {
+
+    }
+}
