+++ app/src/main/java/com/artem/animationjikan/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
 package com.artem.animationjikan.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
